Type category component
